+++ lib/td_auth/permissions/seeds.ex
@@ -74,7 +74,8 @@ defmodule TdAuth.Permissions.Seeds do
       "view_quality_rule",
       "manage_raw_quality_rule_implementations",
       "manage_rule_results",
-      "link_implementation_business_concept"
+      "link_implementation_business_concept",
+      "manage_remediations"
     ],
     "data_sources" => [
       "manage_data_sources"