+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-4271] New permissions `link_implementation_business_concept`
+
 ## [4.40.0] 2022-03-14
 
 ### Changed