+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-4271] New permissions `link_implementation_business_concept`
 - [TD-3233] Rule result remediation plan `manage_remediations` permission
+- [TD-4577] Move view `StructureNotes` permissions to `data_structure` group
 
 ## [4.40.0] 2022-03-14